backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
google-generativeai
python-dotenv
plotly